requirements.txt
# UI & Visualization
streamlit==1.28.0
streamlit-option-menu==0.3.2
streamlit-lottie==0.0.4
streamlit-aggrid==0.3.4
streamlit-extras==0.4.0
streamlit-float==0.3.5
plotly==5.15.0
plotly-calplot==0.1.20
streamlit-plotly-events==0.0.5

# ML & Explainability
scikit-learn>=1.3.0
lightgbm==4.0.0
prophet==1.1.0
shap==0.42.0
matplotlib==3.7.2
scipy>=1.8.0

# Tooling & Quality
python-dotenv==1.0.0
pydantic==2.5.0
joblib==1.3.2

# Data Processing
pandas==2.1.3
numpy==1.24.3
pyyaml==6.0.1

# Reporting
reportlab==4.0.7
python-pptx==0.6.23
weasyprint==60.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Documentation (optional)
mkdocs-material==9.4.8

